fixed interface of casa services, include oauth config for casa services. Include FE Oauth flow to easier config. Cascase Oauth changes to Archetype

+++ casa-service/src/main/java/com/wk/service/CasaService.java
@@ -13,7 +13,7 @@ public class CasaService {
 
     private final CasaAccountRepository repository;
 
-    public List<CasaAccount> getAllByUserId(Integer userId) {
+    public List<CasaAccount> getAllByUserId(Long userId) {
         return repository.getAllCasaByUserId(userId);
     }
 }
